refactor: move MarketingNav and MarketingFooter components to features/home and update imports
